Move location tests to LocationIntegrationTests
* Move location tests from MainOneSignalClassRunner file to LocationIntegrationTests
